No concurrency (#9)
* enh(ssh): auto_ptr replaced by unique_ptr

* enh(ssh): concurrency::mutex replaced by std::mutex

* enh(connector): WIP to suppress concurrency objects from clib

The goal here is to use standard ones.

* enh(concurrency): Standard library used for threads stuff

* fix(test): Failing tests disabled

* cleanup(ssh): debug functions removed.

+++ ssh/inc/com/centreon/connector/ssh/orders/listener.hh
@@ -42,10 +42,10 @@ namespace        orders {
     listener&    operator=(listener const& l);
     virtual void on_eof() = 0;
     virtual void on_error(
-                   unsigned long long cmd_id,
+                   uint64_t cmd_id,
                    char const* msg) = 0;
     virtual void on_execute(
-                   unsigned long long cmd_id,
+                   uint64_t cmd_id,
                    time_t timeout,
                    std::string const& host,
                    unsigned short port,
